making common function for chat bubble and renamed function to minimize confusion
