+++ app/src/main/java/com/papela/hidoop/presentation/chat/components/RegularReceive.kt
@@ -1,0 +1,21 @@
+@file:JvmName("RegularReceiveMessageKt")
+
+package com.papela.hidoop.presentation.chat.components
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.papela.hidoop.presentation.chat.common.SentChatBubbleSurface
+
+@Composable
+fun RegularReceive(content: String) {
+    SentChatBubbleSurface {
+        Text(
+            text = content,
+            modifier = Modifier
+                .padding(16.dp)
+        )
+    }
+}
